feature:前端展示数据传递

+++ common/src/main/java/com/gaokao/common/meta/vo/volunteer/VolunteerFormVO.java
@@ -18,20 +18,28 @@ public class VolunteerFormVO {
 
     private Long id;
 
+    //用户id
     private Long userId;
 
+    //志愿表名称
     private String name;
 
-    private Long score;
+    //用户分数
+    private Integer score;
 
+    //选课信息
     private List<Long> subject;
 
+    //志愿id
     private String volunteerList;
 
+    //生成类型
     private Long generatedType;
 
+    //是否为当前表
     private Long currentIsOrNot;
 
+    //生成时间
     private Long generatedTime;
 
 }
